Refactor unit tests for filesystem driver
Reduced test duplications for single store and multistore cases. Added
missing coverage for thin provisioning in multistore.
